Move storage providers to separate files
